Move `VideoInputOptions` up a directory
