+++ SeeShark/Device/Display.cs
@@ -2,8 +2,6 @@
 // This file is part of SeeShark.
 // SeeShark is licensed under the BSD 3-Clause License. See LICENSE for details.
 
-using SeeShark.Decode;
-
 namespace SeeShark.Device
 {
     public class Display : VideoDevice
